// web/src/api/generated/models/AuthorSummary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AuthorSummary
 */
export interface AuthorSummary {
    /**
     * 
     * @type {string}
     * @memberof AuthorSummary
     */
    displayName: string;
    /**
     * 
     * @type {string}
     * @memberof AuthorSummary
     */
    id: string;
    /**
     * 
     * @type {boolean}
     * @memberof AuthorSummary
     */
    isFormerMember?: boolean;
    /**
     * 
     * @type {string}
     * @memberof AuthorSummary
     */
    profileAttachmentId?: string;
}

/**
 * Check if a given object implements the AuthorSummary interface.
 */
export function instanceOfAuthorSummary(value: object): value is AuthorSummary {
    if (!('displayName' in value) || value['displayName'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    return true;
}

export function AuthorSummaryFromJSON(json: any): AuthorSummary {
    return AuthorSummaryFromJSONTyped(json, false);
}

export function AuthorSummaryFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthorSummary {
    if (json == null) {
        return json;
    }
    return {
        
        'displayName': json['displayName'],
        'id': json['id'],
        'isFormerMember': json['isFormerMember'] == null ? undefined : json['isFormerMember'],
        'profileAttachmentId': json['profileAttachmentId'] == null ? undefined : json['profileAttachmentId'],
    };
}

export function AuthorSummaryToJSON(json: any): AuthorSummary {
    return AuthorSummaryToJSONTyped(json, false);
}

export function AuthorSummaryToJSONTyped(value?: AuthorSummary | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'displayName': value['displayName'],
        'id': value['id'],
        'isFormerMember': value['isFormerMember'],
        'profileAttachmentId': value['profileAttachmentId'],
    };
}